Create file for lesson 3 and add reflections.
i want to see how optional messages work on git bash hence this message, it was optional because here you can provide long messages explaining about the recent commit to be exchanged with the other person you are sharing your code with. now i think the message is long enough to be left alone and check it out in git bash. 

thank you.